refactor(WP05): extract baseline_merge_commit cluster to merge/baseline.py with back-compat aliases (#2027, epic #2026)
